LG-8312 Activate Threatmetrix for all user sessions
This code removes old code that enabled Threatmetrix for a limited set of SPs. This code was in place for testing Threatmetrics. Now that we are confident in the implementation we will be running it on all SPs.

A follow-up commit will be needed to drop database columns and clean up background job arguments.

[skip changelog]