Change the API for `expect_delivered_email` to not include an index
The `expect_delivered_email` helper allows tests to confirm that an email with a given subject/recipient/body was in fact sent.

Prior to this commit this method took an index as the first argument. This meant that emails had to be sent in a specific, deterministic order for tests to pass. Problems arose when a users emails could be returned in a non-deterministic order. In most cases emails should be returned in the order in which they are used to sign in. However, for users created with a factory this may not always be the case.

After reviewing the tests that use this helper it does not appear that we need to actually worry about the order in which the emails are sent for most cases. This commit changes this helper to be agnostic about the order and applies the new API.

[skip changelog]