Adding the warning type to the wrapped Alert
-- What
Preivously we were not wrapping Alert, and therefore had the correct
uswds alert classes manually present on the containing element. Now
that we are wrapping the Alert component we need to set the
appropriate type.

Adding test that ensures the alert is a warning.

changelog: Improvements, Improved Language, Updating hybrid flow doc
capture warning