LG-8312 Drop the device profiling enabled column from service providers
A previous commit stopped reads from this column and ignored it. This commit goes back and cleans it up entirely.

This commit also cleans up some job arguments that we changed in the same commit that stopped reads from this column.

PR that stopped use of this column: https://github.com/18F/identity-idp/pull/7446

[skip changelog]