Avoid ValidatedFieldComponent missing aria-describedby references (#7484)
* Avoid ValidatedFieldComponent missing aria-describedby references

changelog: Bug Fixes, Accessibility, Fix missing ID references for field descriptors

* Fix specs

* Remove unnecessary non-null operator

* Add error-id as attribute of lg-validated-field

Inspired by @svalexander's idea in #7488

Co-Authored-By: Shannon A <20867088+svalexander@users.noreply.github.com>

* Try avoiding layout-affecting padding for empty error message

May want to revisit. Could also consider ":empty" pseudo-selector

* Make spec descriptor hint more realistic

Avoid describedby associated to a label (which would be labelledby)

* Sync MemorableDateComponent markup

See: https://github.com/18F/identity-idp/pull/7484#discussion_r1049085543

* Revert "Try avoiding layout-affecting padding for empty error message"

This reverts commit d12c08afc8bbc9861de68d5b3182cbb2b3b3e15a.

* Update MemorableDate component to add error-id reference

* Create helper methods to consolidate logic for hint/error ID

* Restore display: none styling for valid field

* Add describedby association as part of input validity handling

* Remove accessible description assertion

Previously, I picked "month" as an arbitrary field to find the error, but since now the field validity aria-describedby is managed per-field, the relationship would happen at the field level. We can (and ideally should) check the accessible description for the field we're interested in checking.

* Avoid excess spaces in aria-describedby

technically not an issue, but we should avoid having "empty" id refs be considered

* Assert error message as hidden

we could revert to what existed previously with "not_to have_css", but stronger guarantee that we check both the text is empty, and that it's explicitly present but not visible

* Update MemorableDateComponent specs

* Remove debugging code

Co-authored-by: Shannon A <20867088+svalexander@users.noreply.github.com>