Updating route for pii verify show
-- What
The FSM currently looks up the current step name from a portion of the
current path (whose variable is `doc_auth/:step`. We were previously
hard coding this as `doc_auth/verify_pii`, which then caused the FSM
to look for a step called `verify_pii` which does not exist (it is
simply called `verify`). For this reason, we are reverting to just
`doc_auth/verify` as the path, but still pointing the GET request to
our new `verify_pii` controller's #show method.

[skip changelog]