Updating routes and test for new verify_pii path
-- What
We need to update the verify_step spec to test whether or not we are
explicitly going to the correct show path for the new controller. To
do this, we pass an `as:` of `doc_auth_verify_pii` in routes, which allows us to
know for sure that the right controller is being called in the (now
updated) verify_step spec test case.

[skip changelog]