Updating routes and verify_step spec
-- What
Several key points:
1. We needed to change the actual route name from `verify_pii` to
`verify_info`, because the analytics will check output for strings
containing "pii" as a security measure. (We might consider renaming
the controller as well, to VerifyInfoController. TBD);
2. We needed to add the `verify_info` key to the DocAuthFlow's FSM
steps dictionary, so that the FSM can correctly look up the current
step (which is based on URL attributes);
3. We updated the idv_doc_auth helper to the correct path so that
all specs check that our controller is being called for #show