Remove unused `issuer` arg from `ResolutionProofingJob`
This arg used to be used to determine if ThreatMetrix should be used. This was in place because ThreatMetrix was only active for some service providers. A recent change made Threatmetrix active for all service providers.

This argument was left in place with a `nil` default to prevent ArgumentErrors when the workers were running old and new code during deployment. Now that the code that used this argument is not running in any deployed environment the argument can be safely removed.

[skip changelog]