Add the drivers license number check to the LexisNexis failed attribute mapper
We started sending driver's license numbers to instant verify a while ago. We are using the check as a "warning" indicator currently. The primary purpose of this is to get visibility into how InstantVerify performs for driver's license checks.

The above change means we have a new Item in the products array from the LexisNexis response to handle as a failed attribute. Currently this is labeled as an 'unkown'. Adding this will make it visible as a failure on drivers license number and indicate that attribute requires additional verification. This will cause it to appear in our logs.

[skip changelog]