Retire mailcatcher
After we upgraded to Ruby 3.2 we lost support for mailcatcher. Mailcatcher depends on Fixnum was was previously deprecated and is now removed in Rby 3.2

This commit removes mailcatcher and adds 2 new options for previewing emails:

1. Letter opener
2. File previews

This commit has letter opener as the default method with the ability to switch to the file method by changing configs in your local `application.yml`.

[skip changelog]