Have the verification error parser use the `ProductType` instead of `ExecutedStepName` (#7630)
The LexisNexis InstantVerify response includes a list of products that ran. We parse error messages out of these products and use info from the product as a key for those errors in a hash. Currently we use the `ExecutedStepName` as the key for a product. This key can change between products. This commit makes a change to use the `ProductType` key which should be more consistent as the workflow changes.

[skip changelog]

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>