LG-8574 Start logging document classification information for TrueID (#7646)
In the past we did not log classification information because it reveals information about the applicant's state. We have started logging the state that issued the document elsewhere to provide better debugging. Since the information is already logged we can also start logging this classification info. This will tell us specifically which template the vendor is comparing against. This will allow us to measure the performance of individual templates.

[skip changelog]