Log whether we skipped the upload step when the upload step is submitted
The upload step gets "submitted" regardless of whether or not it is actually visibile to the user (see the `update_if_skipping_upload` before action in the `DocAuthController`). This results in a submission event being logged regardless of whether the user actually submitted the step.

This commit adds a param to the logs which will indicate to use that though there is a submission event the step was skipped an not visible to the user. This will be helpful in building a hybrid flow dashboard to determine how many users actually encounter the hybrid flow selection and what their choices are from that screen.

[skip changelog]