Do not allow people to re-enter the verify info step once it is completed
This is a fix for an issue that is a result of an assumption the flow state machine made. It did not allow steps to be re-entrant without using specific "actions" to ensure the state was correct.

This step adds this behavior to the new verify info controller until the actions it can trigger (i.e. change address and update SSN) can be made to work in a re-entrant manner.

[skip changelog]