Remove the old verify FSM step from the codebase
We are retiring the FSM and the verify step is the first step to be removed. Previous put work in place to retired gracefully. Once this code is no longer running in prod this PR should be safe to merge.

The `VerifyWait` and `VerifyWaitStepShow` were removed as they are associated with the Verify step and are not needed without it.

[skip changelog]