Deleting useless hybrid handoff mobile tests
-- What
Old code was testing for case where user is at the hybrid handoff
screen (upload step), but is on mobile. We have not allowed that to
happen for a long time, so the code is dead.

changelog: User-Facing Improvements, Hybrid Handoff Screen, Combining
upload and send link steps into one screen