LG-8771: Implement reCAPTCHA on phone setup screens (#7761)
* LG-8771: Implement reCAPTCHA on phone setup screens

[skip changelog]

* Render captcha submit button as submit button

No-JS compatibility for feature disabled/exempt cases

* Assign exempt countries only when captcha enabled

By not assigning countries, it falls back to the "always exempt" behavior, which is how it should work when feature is disabled

* Rename exempt to isExempt

Practice what you preach: Inform as boolean

* Render hidden input from CaptchaSubmitButtonComponent

1. Allows a means to show the error message
2. Avoids awkwardness of server-side param permitting
3. Simplifies JavaScript

* Move constants top of file

* Handle Faraday errors for reCAPTCHA validation

* Add logging for reCAPTCHA result

* Move constants top of file

* Add score threshold to analytics

* Add service name to Faraday instrumentation

* Remove default button content for submit button

Needs to be customizable (e.g. "Send one time code")

* Reference submit button by form query

* Consider absence of reCAPTCHA key as exemption

Avoids needing exemption to be considered as external consideration of PhoneInputComponent

* Add README for captcha-submit-button package

* Revert SpinnerButton default action message, default in captcha button

I remember now that it was intentional to avoid the default action message as a way to force the implementer to acknowledge and include a relevant message to be announced to assistive technology. "Verifying..." is pretty widespread, but not really a great choice describing the inherent behavior of a spinner button

* Render captcha submit button as submit button

Default for spinnerbutton is plain button (probably shouldn't be =/)

* Separate configuration check from exemption

"exempt" and "doesn't have requisite config" should be considered as separate things

* Remove unused spinnerButton instance function

* Add specs for CaptchaSubmitButtonElement

* Avoid rendering reCAPTCHA script when reCAPTCHA disabled

* Add specs for CaptchaSubmitButtonComponent

* Normalize YAML

* Stop spinner spinning on invalid form

* Update PhoneSetupController specs

* Short-circuit invalid result for missing reCAPTCHA token

* Remove reCAPTCHA error expectations

Default configuration would have reCAPTCHA disabled, so there ought not be any errors

* Add specs for RecaptchaValidator

* Add specs for PhoneSetupRecaptchaValidator

* Update specs for PhoneInputComponent

* Override CSP for reCAPTCHA only when enabled

* Add specs for RecaptchaConcern

* Update NewPhoneForm specs for recaptcha handling

* Use UriService for adding URL parameters

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>

* Use delegating for PhoneSetupRecaptchaValidator

See https://github.com/18F/identity-idp/pull/7761#discussion_r1099411758

* Remove NewRelic error notice for unsuccessful validation

See: https://github.com/18F/identity-idp/pull/7761#discussion_r1099418836

We are already logging a detail to CloudWatch we can track

* Handle captcha challenge exemption as event cancellation

See: https://github.com/18F/identity-idp/pull/7761#discussion_r1100346531

* Render SpinnerButton explicitly, not as subclass

* Remove class_name from reCAPTCHA validator analytics

We're no longer subclassing, so this wasn't providing an accurate detail. Could consider in the future if we need to disambiguate, at which point maybe we add a "context" property to the class constructor.

* Rename PhoneSetupRecaptcha to PhoneRecaptcha

More abbreviated, could allow reuse, shouldn't be restricted to setup

* Use consistent naming convention for custom events

lg:component:action

* Include evaluated result in analytics payload

* Log class name for reCAPTCHA connection exception

* Add coverage for nil analytics

* Avoid logging for exempt, blank reCAPTCHA

---------

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>