Refactor the IdV step concern to no longer check for an applicant (#7840)
Prior to the introduction of the flow state machine, the "applicant" value in the IdV session was created when the proofing process was started. After the introduction of the FSM the applicant was created once the FSM flow was complete.

The IdvStepConcern is intended to route users to the correct step based on the state of proofing. Currently it is empty because the Flow State Machine primarily handles this logic. There is a desire to re-introduce the logic to route users to the correct step to this concern.

The IdvStepConcern has a before action to verify that proofing has started. It does that by checking for the applicant which, as described above, does not make sense in the post-FSM world. This commit removes that before action and adds it explicitly to controllers that require an applicant to function.

changelog: Improvements, Rails controller management, A before action that was used to identify if a user has started proofing was removed from a concern and explicitly called in individual controllers that depended on that before action. The logic in that before action was superceded by the flow state machine.