Move before actions that confirm steps are done into step concern
The IdV step concern was originally intended to ensure that a users had completed a step before allowing them to advance. This functionality was replaced by the flow state machine.

This commit makes a change to add the logic back for steps that have been removed from FSM or were never a part of FSM.

changelog: Improvements, Flow State Machine, The logic for determining if a user has completed a step outside the context of the flow state machine was consolidated into a controller concern.