LG-8901: New Hybrid Handoff (upload step) with feature flag (#7849)
* Adding initial fields and new icons
* Deleting useless hybrid handoff mobile tests
* Pulling in send_link step functionality to upload_step
* Adding feature flag for new combined hybrid handoff step
* Adding additional checking for 50/50 conditions

-- What
As @matthinz pointed out, the current code does not account well for
the following 50/50 states:
1. User is shown the new combined upload view, but submits the phone
number to an instance with the feature flag off;
2. User is shown the original upload view, but clicks the phone button
submitting to an instance where the feature flag is now on

We handle this case by adding a request param _only_ to the form
rendered by the new combined view. That view is only shown to users
who have the feature flag enabled. The upload step will now check for
that param -- called :combined -- and can determine regardless of the
current state of the feature flag, that the user actually has
submitted a phone number to the upload step itself, and that the send
link step should be skipped.

Co-authored-by: Sonia Connolly <sonia.connolly@gsa.gov>
Co-authored-by: Andrew Duthie <andrew.duthie@gsa.gov>

changelog: User-Facing Improvements, Hybrid Handoff Screen, Combining
upload and send link steps into one screen