LG-9025 Go to new SsnController from hybrid flow when feature flag is set (#7875)
From the DocumentCaptureStep, we only want to redirect to the new SsnController if we are in the DocAuthFlow. This step is reused from the Hybrid Flow, and we don't want to go to the SsnController on mobile. In hybrid flow, we don't return to the DocumentCaptureStep, so redirect to SsnController from LinkSentStep when the feature flag is true as well.

Duplicate the hybrid flow test with doc_auth_ssn_controller_enabled flag set to true and make sure it passes. Note that on my machine even on main without the feature flag I have to set a binding.pry and delay the mobile browser part of the test to get it to pass.

changelog: Internal, refactor, fix hybrid flow with new SsnController (feature flagged)

Co-authored-by: eric-gade <eric.gade@gsa.gov>