Remove the ModuleLength rubocop rule (#7880)
There was an old PR that looks like it intended to remove this rule (ref: https://github.com/18F/identity-idp/pull/5135). It eneded up just moving it.

We have a number of modules that are over this length and I don't think it makes sense to enforce on modules and not other files.

changelog: Internal, Linting rules, The ModuleLength rubocop rule was removed to allow for longer consolidated modules.