LG-9066 Remove the FSM SSN step
We are working on retiring the FSM and have recently put work into moving users to a new controller for entering their SSN. This works follows up on that by removing the old SSN step and the code that directed users their.

changelog: Improvements, FSM Retirements, The old FSM step was removed and supporting code was also removed now that all users should be sent to the newly built SSN controller.