Enforce SSN form submission in the ssn controller (#7905)
Prior to this commit the SSN controller submitted the SSN form but did not take any action on the result if it was unsuccessful. This commit changes the logic so the page is rendered with an error in that case.

Prior to this change the user would fail downstream on resolution. A frontend validation should prevent this from occuring.

changelog: Improvements, FSM Retirement, A bug was fixed where the form that validates the SSN format may not be enforced.