Remove leftover OTP delivery redirect (#7921)
* Remove leftover OTP delivery redirect

We recently consolidated the OTP delivery method controller and the phone controller in IdV. The old OTP delivery method controller is no more. I found a redirect to it and discovered it was still referenced in the routes. This commit cleans up both of those and removes an unused shared example I found in the process.

changelog: Internal, Code cleanup, The following were removed: A redirect to an unused route; tests that tested the redirect to the unused route; a RSpec shared example that included a reference to this unused route but was not actually used in any specs; a template that was previously used by the removed controller; translations for the removed template.