Add `#confirm_verify_info_step_complete` to the IdV step concern
With the retirment of the Flow State Machine we will be using the IdV step concern to confirm that steps are completed throughout the proofing flow.

This commit adds a `#confirm_verify_info_step_complete` which can be used as a before action to confirm verify info is complete. If it is not complete it knows to redirect the user either to the remote verify info or in-person verify info depending on whether the user has a in-person enrollmemnt.

changelog: Improvements, FSM Retirement, A before action that validates that the verify info step was completed was added for use in IdV controllers that render steps that require the verify step to have been completed.