Log LexisNexis errors even on successful transactions
In the past we have not logged errors from the response on passed LexisNexis transaction. The assumption here was that if the transaciton passed there must not be any errors.

Due to nuances in the configuration LexisNexis transacitons can pass if there are tolerable errors or errors that are not deemed risky enough to fail a transaction. These still contain useful information, however.

This commit makes a change to always log the errors in LexisNexis transacitons, even if they are successful.

changelog: Improvements, Error logging, Errors that occur while communicating with LexisNexis Risk Defense Platform services (i.e. InstantVerify and PhoneFinder) are now logged regardless of the verification status of the RDP transaction.