LG-7832: Preserve State ID and Address form data on invalid submissions (#7936)
The changes to `base_flow.rb` help retain the form data when responding to a form submission by permitting reuse of the form (as an instance variable) both in validating the submission and in rendering the new page. The flow state machine instance that keeps a reference to the form steps (and therefore the form) is discarded and recreated on every request.

The updates to `state_id_step.rb` and `address_step.rb` assign the form as an instance variable and merge the session data with form data when rendering the response to an invalid submission. They also permit the removal of misleading reuse of the `doc_auth` symbol for identifying these forms.

* LG-7832: Preserve State ID and Address form data on invalid submissions

changelog: Bug Fixes, In-Person Proofing, Retain State ID and Address on form re-render after invalid submissions

* LG-7832: Fix address test

* LG-7832: Fix state ID step test

* Update app/services/idv/steps/in_person/address_step.rb

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>

* Update app/services/idv/steps/in_person/state_id_step.rb

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>

---------

Co-authored-by: Zach Margolis <zachmargolis@users.noreply.github.com>