Add a `#verify_info_step_complete?` and `mark_verify_info_step_complete!` methods to `Idv::Session` (#7958)
The IdV Session manages the session variables during proofing and is inspected throughout to determine if a user should be on a particular step. This means the session contains a series of flags or artifacts that determine if steps were completed. The names of these and the impacts of changing them may not be obvious, so in this commit some methods for marking a step complete and checking that a step is complete were added alongside methods for invalidating a step.

This is a pattern we intend to apply to other steps in the proofing flow.

changelog: Improvement, FSM Retirement, Methods were added to the Idv Session class for determining whether the verify info step was completed and for marking the verify info step complete.