Clean up `ResolutionProofingJob` tests
The `ResolutionProofingJob` is the class that does much of the legwork during the verify step. Specifically, it does the following:

1. Communicates with ThreatMetrix if necessary
2. Communicates with InstantVerify
3. Communicates with AAMVA
4. Combines results through the `ResolutionResultAdjudicator`

The tests for this class have been sophisticated and have combined a number of different ways of mocking out responses from vendors (e.g. webmocks or stubs of the vendor classes).

This change attempts to simplify some of this by creating a helper for stubbing responses from vendors. Those helpers are used to re-implement simpler examples for the logic the tests test. Ideally this will allow us to more quickly make changes to the `ResolutionProofingJob`.

[skip changelog]