Don't make a request to AAMVA if verification is guaranteed to fail (#7997)
* Don't make a request to AAMVA if verification is guaranteed to fail

We implemented a feature that allowed us to use the response we get from AAMVA to cover attributes that fail the LexisNexis check during resolution. Specifically, the following attributes can be covered by AAMVA:

- Address
- Dob
- State ID Number

As part of this change we started sending requests to AAMVA even if the LexisNexis response failed. Previously this occurred even if the failed LexisNexis attributes could not be covered by the AAMVA response.

This commit attempts to shed some load on AAMVA's service by not making requests to them if the attributes in a failed LexisNexis response cannot be covered by AAMVA.

[skip changelog]