Revert toNextStep for IPP step
This should probably be available as an option for a step, but it doesn't work in IPP because IPP relies on frame ticks resulting from the click on the button, since otherwise the state change in setting submit metadata results in the subsequent submit event using a stale version of that metadata state. In earlier iterations here, we changed this to a mutative ref, but refs are not good for use in the public interface of the hook, since updating the ref would not trigger an update to consuming components.