Updating to StatusPageComponent and fixing Exit button
-- What
1. Updated the Outage show view to use the StatusPageComponent, for
better uniformity;
2. Added a controller helper to determine which url to link to in the
"Exit" button. In the case where there is an SP present and it has a
valid return_to_sp_url set, the exit button will navigate to that
URL. In all other cases it will return the user to the account page.

NOTE: We had to change the yml configuration for local testing SPs so
that it had a correct return_to_sp_url that we could actually test
against.

Co-authored-by: Douglas Price <douglas.price@gsa.gov>