Conditionally remove SendLink step from DocAuthFlow (#8033)
* Temporary! Conditionally remove SendLink step from DocAuthFlow

When the doc_auth_combined_hybrid_handoff_enabled feature flag is enabled, the SendLink step is not used. Conditionally remove it from the DocAuthFlow so that the send_link url is not accessed in the 50/50 state when the feature flag and SendLink code are fully removed.

[skip changelog]

* Fix UploadStep step spec

UploadStep spec was testing SendLinkStep instead. Update to test the right step.

Co-authored-by: Eric Gade <eric.gade@gsa.gov>
Co-authored-by: Kimball Bighorse <kimball.bighorse@gsa.gov>