Replace resolution result DLN info with a DLN match boolean (#8005)
Previously, we added a check to log the drivers license check status that we get back from LexisNexis InstantVerify. This check status is present on some workflows. This check requires us to send the driver's license number (DLN) to InstantVerify to pass.

We added tooling to the resolution proofer that logged the DLN check result. This enabled us to make projections about how using this check into proofing flow would affect performance.

While looking at the result from this check we learned that are actually 2 driver's license checks. We should require both to pass to consider the transaction a success.

This commit reconfigures the way that we inspect the LexisNexis InstantVerify driver's license check. It adds a boolean that is true if the check passes. This takes into account the following:

- Both drivers license checks must pass to consider the drivers license a match
- Both drivers license checks must be present to consider the drivers license a match

A previous change started logging all LexisNexis statuses during proofing regardless of the transaction status. This means we are logging the drivers license check status elsewhere. As a result we do not need to worry about losing visibility into the more granular results of the drivers license check.

[skip changelog]