LG-9301 capture error and alert when ArcGIS service is not returning a token
If ArcGIS is unable to return a token it was causing an exception in retrieve_token
and throwing HTTP 500 responses. The UI had no indication of error and the
responses were causing the on-call engineer to be notified.
This change catches the missing token, logs the issue, and returns a 400 error which
allows a retry.

changelog: Internal, PO-Search, Return 400 instead of 500 and logs ArcGIS token issues