LG-9065 Write an exception result when TMx has an invalid review status
We use ThreatMetrix as a tool for device profiling. It gives us a `review_status` result which we store in the database as a proofing component. That proofing component is used later in the process to determine if a profile needs additional review.

We communicate with ThreatMetrix when we are running the `ResolutionProofingJob` if device profiling is enabled or in collect only mode. We do not communicate with threatMetrix when device profiling is disabled.

Previously, we would write `nil` to the proofing component if we received that or an unexpected review status from ThreatMetrix. This leads to ambiguity later in the process when determining if a user passed the ThreatMetrix check. Specifically, `nil` could mean the ThreatMetrix did not run _or_ it could mean that an error occured and no `review_status` was provided.

This commit changes the behavior of ThreatMetrix such that if ThreatMetrix is enabled it will write a review status or the job will respond with an exception result. This exception result will result in the user seeing an error and needing to re-submit. This resolves the ambiguity by making `nil` mean that ThreatMetrix was not enabled.

[skip changelog]